+++ components/config/src/config.rs
@@ -12,6 +12,7 @@ use crate::highlighting::THEME_SET;
 use crate::theme::Theme;
 use errors::{bail, Error, Result};
 use utils::fs::read_file_with_error;
+use utils::slugs::SlugifyStrategy;
 
 // We want a default base url for tests
 static DEFAULT_BASE_URL: &str = "http://a-website.com";
@@ -21,6 +22,24 @@ pub enum Mode {
     Build,
     Serve,
     Check,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(default)]
+pub struct Slugify {
+    pub paths: SlugifyStrategy,
+    pub taxonomies: SlugifyStrategy,
+    pub anchors: SlugifyStrategy,
+}
+
+impl Default for Slugify {
+    fn default() -> Self {
+        Slugify {
+            paths: SlugifyStrategy::On,
+            taxonomies: SlugifyStrategy::On,
+            anchors: SlugifyStrategy::On,
+        }
+    }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -35,7 +54,7 @@ pub struct Language {
 }
 
 impl Default for Language {
-    fn default() -> Language {
+    fn default() -> Self {
         Language { code: String::new(), rss: false, search: false }
     }
 }
@@ -75,7 +94,7 @@ impl Taxonomy {
 }
 
 impl Default for Taxonomy {
-    fn default() -> Taxonomy {
+    fn default() -> Self {
         Taxonomy {
             name: String::new(),
             paginate_by: None,
@@ -130,8 +149,6 @@ pub struct Config {
     /// key into different language.
     translations: HashMap<String, TranslateTerm>,
 
-    /// Whether to slugify page and taxonomy URLs (disable for UTF-8 URLs)
-    pub slugify_paths: bool,
     /// Whether to highlight all code blocks found in markdown files. Defaults to false
     pub highlight_code: bool,
     /// Which themes to use for code highlighting. See Readme for supported themes
@@ -170,6 +187,9 @@ pub struct Config {
     pub extra_syntax_set: Option<SyntaxSet>,
 
     pub link_checker: LinkChecker,
+
+    /// The setup for which slugification strategies to use for paths, taxonomies and anchors
+    pub slugify: Slugify,
 
     /// All user params set in [extra] in the config
     pub extra: HashMap<String, Toml>,
@@ -356,7 +376,6 @@ impl Default for Config {
             title: None,
             description: None,
             theme: None,
-            slugify_paths: true,
             highlight_code: false,
             highlight_theme: "base16-ocean-dark".to_string(),
             default_language: "en".to_string(),
@@ -374,6 +393,7 @@ impl Default for Config {
             extra_syntaxes: Vec::new(),
             extra_syntax_set: None,
             link_checker: LinkChecker::default(),
+            slugify: Slugify::default(),
             extra: HashMap::new(),
             build_timestamp: Some(1),
         }
@@ -382,7 +402,7 @@ impl Default for Config {
 
 #[cfg(test)]
 mod tests {
-    use super::{Config, Theme};
+    use super::{Config, SlugifyStrategy, Theme};
 
     #[test]
     fn can_import_valid_config() {
@@ -618,5 +638,23 @@ skip_prefixes = [
             config.link_checker.skip_prefixes,
             vec!["http://[2001:db8::]/", "https://www.example.com/path",]
         );
+    }
+
+    #[test]
+    fn slugify_strategies() {
+        let config_str = r#"
+title = "My site"
+base_url = "example.com"
+
+[slugify]
+paths = "on"
+taxonomies = "safe"
+anchors = "off"
+        "#;
+
+        let config = Config::parse(config_str).unwrap();
+        assert_eq!(config.slugify.paths, SlugifyStrategy::On);
+        assert_eq!(config.slugify.taxonomies, SlugifyStrategy::Safe);
+        assert_eq!(config.slugify.anchors, SlugifyStrategy::Off);
     }
 }