+++ components/rendering/tests/markdown.rs
@@ -6,6 +6,7 @@ use config::Config;
 use front_matter::InsertAnchor;
 use rendering::{render_content, RenderContext};
 use templates::ZOLA_TERA;
+use utils::slugs::SlugifyStrategy;
 
 #[test]
 fn can_do_render_content_simple() {
@@ -350,7 +351,7 @@ fn can_add_non_slug_id_to_headings() {
     let tera_ctx = Tera::default();
     let permalinks_ctx = HashMap::new();
     let mut config = Config::default();
-    config.slugify_paths = false;
+    config.slugify.anchors = SlugifyStrategy::Safe;
     let context = RenderContext::new(&tera_ctx, &config, "", &permalinks_ctx, InsertAnchor::None);
     let res = render_content(r#"# L'écologie et vous"#, &context).unwrap();
     assert_eq!(res.body, "<h1 id=\"L'écologie_et_vous\">L'écologie et vous</h1>\n");