+++ components/utils/src/slugs.rs
@@ -1,3 +1,16 @@
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum SlugifyStrategy {
+    /// Classic slugification, the default
+    On,
+    /// No slugification, only remove unsafe characters for filepaths/urls
+    Safe,
+    /// Nothing is changed, hope for the best!
+    Off,
+}
+
 fn strip_chars(s: &str, chars: &str) -> String {
     let mut sanitized_string = s.to_string();
     sanitized_string.retain(|c| !chars.contains(c));
@@ -6,37 +19,23 @@ fn strip_chars(s: &str, chars: &str) -> String {
 
 fn strip_invalid_paths_chars(s: &str) -> String {
     // NTFS forbidden characters : https://gist.github.com/doctaphred/d01d05291546186941e1b7ddc02034d3
-    // Also we need to trim . from the end of filename
+    // Also we need to trim whitespaces and `.` from the end of filename
     let trimmed = s.trim_end_matches(|c| c == ' ' || c == '.');
-    let cleaned = trimmed.replace(" ", "_");
-    // And () [] since they are not allowed in markdown links
-    strip_chars(&cleaned, "<>:/|?*#()[]\n\"\\\r\t")
+    strip_chars(&trimmed, r#"<>:"/\|?*"#)
 }
 
-fn strip_invalid_anchors_chars(s: &str) -> String {
-    // spaces are not valid in markdown links
-    let cleaned = s.replace(" ", "_");
-    // https://tools.ietf.org/html/rfc3986#section-3.5
-    strip_chars(&cleaned, "\"#%<>[\\]()^`{|}")
-}
-
-pub fn maybe_slugify_paths(s: &str, slugify: bool) -> String {
-    if slugify {
-        // ASCII slugification
-        slug::slugify(s)
-    } else {
-        // Only remove forbidden characters
-        strip_invalid_paths_chars(s)
+pub fn slugify_paths(s: &str, strategy: SlugifyStrategy) -> String {
+    match strategy {
+        SlugifyStrategy::On => slug::slugify(s),
+        SlugifyStrategy::Safe => strip_invalid_paths_chars(s),
+        SlugifyStrategy::Off => s.to_string(),
     }
 }
 
-pub fn maybe_slugify_anchors(s: &str, slugify: bool) -> String {
-    if slugify {
-        // ASCII slugification
-        slug::slugify(s)
-    } else {
-        // Only remove forbidden characters
-        strip_invalid_anchors_chars(s)
+pub fn slugify_anchors(s: &str, strategy: SlugifyStrategy) -> String {
+    match strategy {
+        SlugifyStrategy::On => slug::slugify(s),
+        SlugifyStrategy::Safe | SlugifyStrategy::Off => s.replace(|c: char| c.is_ascii_whitespace(), "_"),
     }
 }
 
@@ -45,61 +44,44 @@ mod tests {
     use super::*;
 
     #[test]
-    fn strip_invalid_paths_chars_works() {
+    fn can_slugify_paths() {
         let tests = vec![
-            // no newlines
-            ("test\ntest", "testtest"),
-            // no whitespaces
-            ("test ", "test"),
-            ("t est ", "t_est"),
-            // invalid NTFS
-            ("test .", "test"),
-            ("test. ", "test"),
-            ("test#test/test?test", "testtesttesttest"),
-            // Invalid CommonMark chars in links
-            ("test (hey)", "test_hey"),
-            ("test (hey", "test_hey"),
-            ("test hey)", "test_hey"),
-            ("test [hey]", "test_hey"),
-            ("test [hey", "test_hey"),
-            ("test hey]", "test_hey"),
-            // UTF-8
-            ("日本", "日本"),
+            // input, (on, safe, off)
+            ("input", ("input", "input", "input")),
+            ("test ", ("test", "test", "test ")),
+            ("tes t", ("tes-t", "tes t", "tes t")),
+            // Invalid NTFS
+            ("dot. ", ("dot", "dot", "dot. ")),
+            ("日本", ("ri-ben", "日本", "日本")),
+            ("héhé", ("hehe", "héhé", "héhé")),
+            ("test (hey)", ("test-hey", "test (hey)", "test (hey)")),
         ];
 
-        for (input, expected) in tests {
-            assert_eq!(strip_invalid_paths_chars(&input), expected);
+        for (input, (on, safe, off)) in tests {
+            assert_eq!(on, slugify_paths(input, SlugifyStrategy::On));
+            assert_eq!(safe, slugify_paths(input, SlugifyStrategy::Safe));
+            assert_eq!(off, slugify_paths(input, SlugifyStrategy::Off));
         }
     }
 
     #[test]
-    fn strip_invalid_anchors_chars_works() {
+    fn can_slugify_anchors() {
         let tests = vec![
-            ("日本", "日本"),
-            // Some invalid chars get removed
-            ("test#", "test"),
-            ("test<", "test"),
-            ("test%", "test"),
-            ("test^", "test"),
-            ("test{", "test"),
-            ("test|", "test"),
-            ("test(", "test"),
-            // Spaces are replaced by `_`
-            ("test hey", "test_hey"),
+            // input, (on, safe, off)
+            ("input", ("input", "input", "input")),
+            ("test ", ("test", "test_", "test_")),
+            ("tes t", ("tes-t", "tes_t", "tes_t")),
+            // Invalid NTFS
+            ("dot. ", ("dot", "dot._", "dot._")),
+            ("日本", ("ri-ben", "日本", "日本")),
+            ("héhé", ("hehe", "héhé", "héhé")),
+            ("test (hey)", ("test-hey", "test_(hey)", "test_(hey)")),
         ];
 
-        for (input, expected) in tests {
-            assert_eq!(strip_invalid_anchors_chars(&input), expected);
+        for (input, (on, safe, off)) in tests {
+            assert_eq!(on, slugify_anchors(input, SlugifyStrategy::On));
+            assert_eq!(safe, slugify_anchors(input, SlugifyStrategy::Safe));
+            assert_eq!(off, slugify_anchors(input, SlugifyStrategy::Off));
         }
-    }
-
-    #[test]
-    fn maybe_slugify_paths_enabled() {
-        assert_eq!(maybe_slugify_paths("héhé", true), "hehe");
-    }
-
-    #[test]
-    fn maybe_slugify_paths_disabled() {
-        assert_eq!(maybe_slugify_paths("héhé", false), "héhé");
     }
 }